+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - ids parameter to Transaction.query
+- ids parameter to Transfer.query
 
 ## [0.9.0] - 2020-08-12
 ### Added